+++ workspace/examples/src/as-library/simple.ts
@@ -1,12 +1,12 @@
-import * as PT from '@creative-introvert/prediction-testing';
-import {Effect, Option} from 'effect';
+import {randomUUID} from 'node:crypto';
 
-type TestCase = PT.Test.TestCase<
-    {BRAND: string; MODEL?: number; MACHINE_TYPE?: string},
-    {BRAND: string; MODEL?: number}
->;
+import * as PT from '@creative-introvert/prediction-testing';
+import {Console, Effect, Layer, Option, Stream} from 'effect';
+
+type TestCase = PT.Test.TestCase<unknown, {BRAND: string; MODEL?: number}>;
 
 const testCases: TestCase[] = [
+    {input: null, expected: {BRAND: 'Claas'}},
     {input: {BRAND: 'Claas'}, expected: {BRAND: 'Claas'}},
     {input: {BRAND: 'John Deere'}, expected: {BRAND: 'John Deere'}},
     {
@@ -23,32 +23,63 @@ const testCases: TestCase[] = [
             BRAND: 'John Deere',
             MODEL: 8400,
         },
+        tags: ['tractor', 'asdf'],
     },
 ];
 
-const testRun1 = await PT.Test.runAll({
-    testCases,
-    program: input => Effect.sync(() => input),
+const main = Effect.gen(function* () {
+    const testRespository = yield* PT.TestRepository.TestRepository;
+    const name = 'cli-simple';
+
+    // yield* testRespository.commitCurrentTestRun({
+    //     name,
+    //     hash: randomUUID(),
+    // });
+
+    // yield* testRespository.clearCurrentTestRun(name);
+    const currentTestRun =
+        yield* testRespository.getOrCreateCurrentTestRun(name);
+    // yield* testRespository.purgeStaleTestRuns(name);
+
+    // const trs = yield* PT.Test.all({
+    //     testCases,
+    //     program: input => Effect.sync(() => input),
+    //     name,
+    // }).pipe(Stream.runCollect);
+
+    yield* Console.log(yield* testRespository.getAllTestRuns);
+    // yield* Console.log(
+    //     yield* testRespository
+    //         .getTestResultsStream(currentTestRun)
+    //         .pipe(Stream.runCollect),
+    // );
+    yield* Console.log(yield* testRespository.getAllTestResults);
+
+    // const testRun2 = await PT.Test.runAll({
+    //     testCases,
+    //     program: input =>
+    //         Effect.sync(() => ({...input, MODEL: input?.MODEL ?? 0 + 100})),
+    // });
+    // console.log(
+    //     PT.Show.summarize({
+    //         testRun: testRun1,
+    //         previousTestRun: Option.some(testRun2),
+    //     }),
+    // );
+    //
+    // console.log(PT.Show.stats({testRun: testRun1}));
+    //
+    // const diff = PT.Test.diff({
+    //     testRun: testRun1,
+    //     previousTestRun: Option.some(testRun2),
+    // });
+    //
+    // console.log(PT.Show.diff({diff}));
 });
 
-const testRun2 = await PT.Test.runAll({
-    testCases,
-    program: input =>
-        Effect.sync(() => ({...input, MODEL: input?.MODEL ?? 0 + 100})),
-});
-
-console.log(
-    PT.Show.summarize({
-        testRun: testRun1,
-        previousTestRun: Option.some(testRun2),
-    }),
+const LiveLayer = Layer.provideMerge(
+    PT.TestRepository.LiveLayer,
+    PT.TestRepository.makeSqliteLiveLayer('as-library-simple.db'),
 );
 
-console.log(PT.Show.stats({testRun: testRun1}));
-
-const diff = PT.Test.diff({
-    testRun: testRun1,
-    previousTestRun: Option.some(testRun2),
-});
-
-console.log(PT.Show.diff({diff}));
+void main.pipe(Effect.provide(LiveLayer), Effect.runPromise);