+++ workspace/prediction-testing-cli/src/Config.ts
@@ -1,0 +1,14 @@
+import type * as PT from '@creative-introvert/prediction-testing';
+
+import * as P from './prelude.js';
+
+export type Config<I = unknown, O = unknown, T = unknown> = {
+    testSuite: PT.Test.TestSuite<I, O, T>;
+    dbPath: string;
+    displayConfig?: Partial<PT.DisplayConfig.DisplayConfig> | undefined;
+};
+
+export const Config = P.Context.GenericTag<Config>('Config');
+
+export const makeConfigLayer = (config: Config) =>
+    P.Layer.sync(Config, () => Config.of(config));