+++ workspace/examples/CHANGELOG.md
@@ -5,6 +5,22 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @creative-introvert/prediction-testing-cli@0.2.1
+  - @creative-introvert/prediction-testing@0.2.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @creative-introvert/prediction-testing-cli@0.2.0
+  - @creative-introvert/prediction-testing@0.2.0
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
   - @creative-introvert/prediction-testing@0.1.0
   - @creative-introvert/prediction-testing-cli@0.1.0